:recycle: Refactor slash commands and add global commands (#4)
